+++ WootricSDK/WootricSDK/WTRApiClient.m
@@ -2,7 +2,7 @@
 //  WTRApiClient.m
 //  WootricSDK
 //
-// Copyright (c) 2015 Wootric (https://wootric.com)
+// Copyright (c) 2018 Wootric (https://wootric.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -24,6 +24,7 @@
 
 #import "WTRApiClient.h"
 #import "WTRPropertiesParser.h"
+#import "WTRLogger.h"
 #import <CommonCrypto/CommonHMAC.h>
 
 @interface WTRApiClient ()
@@ -97,7 +98,7 @@
   NSMutableURLRequest *urlRequest = [self requestWithURL:url HTTPMethod:nil andHTTPBody:nil];
   NSURLSessionDataTask *dataTask = [_wootricSession dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
     if (error) {
-      NSLog(@"WootricSDK (GET end user): %@", error);
+      [WTRLogger logError:@"(GET end user): %@", error];
     } else {
       id responseJSON = [NSJSONSerialization JSONObjectWithData:data options:0 error:nil];
       if ([responseJSON isKindOfClass:[NSArray class]]) {
@@ -110,14 +111,14 @@
           
           if (endUser[@"id"]) {
             NSInteger endUserID = [endUser[@"id"] integerValue];
-            if (!_endUserAlreadyUpdated) {
+            if (!self.endUserAlreadyUpdated) {
               [self updateExistingEndUser:endUserID];
             }
             endUserWithID(endUserID);
           }
         }
       } else {
-        NSLog(@"WootricSDK - Error: %@", responseJSON);
+        [WTRLogger logError:@"%@", responseJSON];
       }
     }
   }];
@@ -159,10 +160,10 @@
 
     NSURLSessionDataTask *dataTask = [_wootricSession dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
       if (error) {
-        NSLog(@"WootricSDK (update end user): %@", error);
+        [WTRLogger logError:@"(update end user): %@", error];
       } else {
-        NSLog(@"WootricSDK (update end user): user updated");
-        _endUserAlreadyUpdated = YES;
+        [WTRLogger log:@"(update end user): user updated"];
+        self->_endUserAlreadyUpdated = YES;
       }
     }];
     
@@ -200,10 +201,10 @@
 
   NSURLSessionDataTask *dataTask = [_wootricSession dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
     if (error) {
-      NSLog(@"WootricSDK (create end user): %@", error);
+      [WTRLogger logError:@"(create end user): %@", error];
     } else {
       NSDictionary *responseJSON = [NSJSONSerialization JSONObjectWithData:data options:0 error:nil];
-      NSLog(@"WootricSDK (create end user): %@", responseJSON);
+      [WTRLogger log:@"(create end user): %@", responseJSON];
       if (responseJSON) {
         NSInteger endUserID = [responseJSON[@"id"] integerValue];
         endUserWithID(endUserID);
@@ -222,9 +223,9 @@
   
   NSURLSessionDataTask *dataTask = [_wootricSession dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
     if (error) {
-      NSLog(@"ResponseError: %@", error);
+      [WTRLogger logError:@"ResponseError: %@", error];
     } else {
-      NSLog(@"Create response added to queue");
+      [WTRLogger log:@"Create response added to queue"];
     }
   }];
 
@@ -244,14 +245,14 @@
 
   NSURLSessionDataTask *dataTask = [_wootricSession dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
     if (error) {
-      NSLog(@"WootricSDK (authentication): %@", error);
+      [WTRLogger logError:@"(authentication): %@", error];
     } else {
       NSDictionary *responseJSON = [NSJSONSerialization JSONObjectWithData:data options:0 error:nil];
-      NSLog(@"WootricSDK (authentication): %@", responseJSON);
+      [WTRLogger log:@"(authentication): %@", responseJSON];
       if (responseJSON) {
         NSString *accessToken = responseJSON[@"access_token"];
         if (accessToken) {
-          _accessToken = accessToken;
+          self->_accessToken = accessToken;
           authenticated();
         }
       }
@@ -278,36 +279,40 @@
   NSMutableURLRequest *urlRequest = [NSMutableURLRequest requestWithURL:url];
   [urlRequest setValue:@"Wootric-Mobile-SDK" forHTTPHeaderField:@"User-Agent"];
   
-  NSLog(@"WootricSDK: eligibility - %@", urlRequest);
+  [WTRLogger log:@"eligibility - %@", urlRequest];
 
   NSURLSessionDataTask *dataTask = [_wootricSession dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
     if (error) {
-      NSLog(@"WootricSDK: %@", error);
+      [WTRLogger logError:@"%@", error];
     } else {
       NSDictionary *responseJSON = [NSJSONSerialization JSONObjectWithData:data options:0 error:nil];
       if (responseJSON) {
         if ([responseJSON[@"eligible"] isEqual:@1]) {
-          if (_settings.forceSurvey) NSLog(@"WootricSDK: forced survey (remove for production!)");
-          NSLog(@"WootricSDK: User eligible");
+          if (self->_settings.forceSurvey) {
+            [WTRLogger logError:@"forced survey (remove for production!)"];
+          }
+          [WTRLogger log:@"User eligible"];
 
-          [_settings parseDataFromSurveyServer:responseJSON];
-          _userID = responseJSON[@"settings"][@"user_id"];
+          [self->_settings parseDataFromSurveyServer:responseJSON];
+          self->_userID = responseJSON[@"settings"][@"user_id"];
           
-          _uniqueLink = [self buildUniqueLinkAccountToken:_accountToken
-                                             endUserEmail:[_settings getEndUserEmailOrUnknown]
+          self->_uniqueLink = [self buildUniqueLinkAccountToken:self->_accountToken
+                                             endUserEmail:[self->_settings getEndUserEmailOrUnknown]
                                                      date:[[NSDate date] timeIntervalSince1970]
                                              randomString:[self randomString]];
           
           if (responseJSON[@"settings"][@"account_id"] != nil) {
-            _accountID = responseJSON[@"settings"][@"account_id"];
+            self->_accountID = responseJSON[@"settings"][@"account_id"];
           }
           eligible();
         } else {
-          NSString *logString = @"WootricSDK: User ineligible";
+          NSString *logString = @"User ineligible";
           if (responseJSON[@"error"]){
             logString = [NSString stringWithFormat:@"%@ - %@", logString, responseJSON[@"error"]];
+            [WTRLogger logError:@"%@", logString];
+          } else {
+            [WTRLogger log:@"%@", logString];
           }
-          NSLog(@"%@", logString);
         }
       }
     }