+++ WootricSDK/WootricSDK/WTRSurvey.m
@@ -2,7 +2,7 @@
 //  WTRSurvey.m
 //  WootricSDK
 //
-// Copyright (c) 2015 Wootric (https://wootric.com)
+// Copyright (c) 2018 Wootric (https://wootric.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -25,6 +25,7 @@
 #import "WTRSurvey.h"
 #import "WTRApiClient.h"
 #import "WTRDefaults.h"
+#import "WTRLogger.h"
 
 @interface WTRSurvey ()
 
@@ -58,7 +59,7 @@
   
   if (_apiClient.settings.forceSurvey || [self needsSurvey]) {
     [_apiClient checkEligibility:^{
-      [_apiClient authenticate:^{
+      [self.apiClient authenticate:^{
           showSurvey();
       }];
     }];
@@ -76,32 +77,32 @@
     } else {
       days = _apiClient.settings.surveyedDefaultDurationDecline;
     }
-    NSLog(@"WootricSDK: needsSurvey(NO) - Already surveyed in last %li days", (long)days);
+    [WTRLogger log:@"needsSurvey(NO) - Already surveyed in last %li days", (long)days];
     return NO;
   } else if (_apiClient.settings.surveyImmediately) {
-    NSLog(@"WootricSDK: needsSurvey(YES) - surveyImmediately");
+    [WTRLogger log:@"needsSurvey(YES) - surveyImmediately"];
     return YES;
   } else if (!_apiClient.settings.externalCreatedAt) {
-    NSLog(@"WootricSDK: needsSurvey(YES) - no externalCreatedAt");
+    [WTRLogger log:@"needsSurvey(YES) - no externalCreatedAt"];
     return YES;
   } else {
     if ([_apiClient.settings.firstSurveyAfter intValue] > 0) {
       NSInteger age = [[NSDate date] timeIntervalSince1970] - [_apiClient.settings.externalCreatedAt intValue];
       if (age > ([_apiClient.settings.firstSurveyAfter intValue] * 60 * 60 * 24)) {
-        NSLog(@"WootricSDK: needsSurvey(YES) - end user's account older than firstSurveyAfter value");
+        [WTRLogger log:@"needsSurvey(YES) - end user's account older than firstSurveyAfter value"];
         return YES;
       } else {
         if (([[NSDate date] timeIntervalSince1970] - [defaults doubleForKey:@"lastSeenAt"]) >= ([_apiClient.settings.firstSurveyAfter intValue] * 60 * 60 * 24)) {
-          NSLog(@"WootricSDK: needsSurvey(YES) - end user's lastSeenAt greater than or equal firstSurveyAfter value");
+          [WTRLogger log:@"needsSurvey(YES) - end user's lastSeenAt greater than or equal firstSurveyAfter value"];
           return YES;
         }
       }
     } else {
-      NSLog(@"WootricSDK: needsSurvey(YES) - firstSurveyAfter is set to less than or equal 0 in SDK (will be compared to admin panel value)");
+      [WTRLogger log:@"needsSurvey(YES) - firstSurveyAfter is set to less than or equal 0 in SDK (will be compared to admin panel value)"];
       return YES;
     }
   }
-  NSLog(@"WootricSDK: needsSurvey(NO) - No check passed");
+  [WTRLogger log:@"needsSurvey(NO) - No check passed"];
   return NO;
 }