+++ WootricSDK/WootricSDK/WTRiPADQuestionView.m
@@ -2,7 +2,7 @@
 //  WTRiPADQuestionView.m
 //  WootricSDK
 //
-// Copyright (c) 2015 Wootric (https://wootric.com)
+// Copyright (c) 2018 Wootric (https://wootric.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal