+++ WootricSDK/WootricSDK/Wootric.m
@@ -2,7 +2,7 @@
 //  WootricSDK.m
 //  WootricSDK
 //
-// Copyright (c) 2015 Wootric (https://wootric.com)
+// Copyright (c) 2018 Wootric (https://wootric.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -28,6 +28,7 @@
 #import "WTRSurveyViewController.h"
 #import "WTRiPADSurveyViewController.h"
 #import "WTRApiClient.h"
+#import "WTRLogger.h"
 
 #define IPAD UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad
 
@@ -145,7 +146,8 @@
     WTRSurvey *surveyClient = [[WTRSurvey alloc] init];
     [surveyClient survey:^{
         
-      NSLog(@"WootricSDK: presenting survey view");
+      [WTRLogger log:@"presenting survey view"];
+
       
       WTRApiClient *apiClient = [WTRApiClient sharedInstance];
       
@@ -158,7 +160,7 @@
         
     }];
   } else {
-    NSLog(@"WootricSDK: Configure SDK first");
+    [WTRLogger log:@"Configure SDK first"];
   }
 }
 
@@ -275,6 +277,20 @@
 + (void)setThankYouButtonBackgroundColor:(UIColor *)color {
   WTRApiClient *apiClient = [WTRApiClient sharedInstance];
   [apiClient.settings setThankYouButtonBackgroundColor:color];
+}
+
+#pragma mark - WTRLogger setters
+
++ (void)setLogLevelNone {
+  [WTRLogger setLogLevel:WTRLogLevelNone];
+}
+
++ (void)setLogLevelError {
+  [WTRLogger setLogLevel:WTRLogLevelError];
+}
+
++ (void)setLogLevelVerbose {
+  [WTRLogger setLogLevel:WTRLogLevelVerbose];
 }
 
 @end